+++ src/MDC/DataCollector.cpp
@@ -1,0 +1,4 @@
+#include "DataCollector.h"
+
+DataCollector::DataCollector(Model* model) : model_(model) {}
+